@@ -1,11 +1,11 @@
 name: google_map_location_picker
 description: '🌍 Map location picker for flutter Based on google_maps_flutter'
-version: 4.1.7
+version: 4.1.8
 homepage: https://github.com/humazed/google_map_location_picker
 
 #  flutter pub pub publish --dry-run
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=2.0.0"
 
 dependencies:
@@ -14,15 +14,16 @@ dependencies:
 
   flutter_localizations:
     sdk: flutter
-  intl: ^0.17.0
+  intl: ^0.18.0
 
-  google_maps_flutter: ^2.0.1
-  geolocator: ^6.1.14
-  android_intent: ^2.0.0
-  provider: ^5.0.0
-  http: ^0.13.0
+  google_maps_flutter: ^2.0.6
+  geolocator: ^9.0.2
+  android_intent: ^2.0.2
+  provider: ^6.0.4
+  http: ^0.13.3
   stack_trace: ^1.10.0
-  package_info: ^2.0.0
+  package_info: ^2.0.2
+
 
 dev_dependencies:
   flutter_test:
