@@ -22,7 +22,7 @@ import 'utils/location_utils.dart';
 class LocationPicker extends StatefulWidget {
   LocationPicker(
     this.apiKey, {
-    Key key,
+    Key? key,
     this.initialCenter,
     this.initialZoom,
     this.requiredGPS,
@@ -44,28 +44,28 @@ class LocationPicker extends StatefulWidget {
 
   final String apiKey;
 
-  final LatLng initialCenter;
-  final double initialZoom;
-  final List<String> countries;
+  final LatLng? initialCenter;
+  final double? initialZoom;
+  final List<String>? countries;
 
-  final bool requiredGPS;
-  final bool myLocationButtonEnabled;
-  final bool layersButtonEnabled;
-  final bool automaticallyAnimateToCurrentLocation;
+  final bool? requiredGPS;
+  final bool? myLocationButtonEnabled;
+  final bool? layersButtonEnabled;
+  final bool? automaticallyAnimateToCurrentLocation;
 
-  final String mapStylePath;
+  final String? mapStylePath;
 
-  final Color appBarColor;
-  final BoxDecoration searchBarBoxDecoration;
-  final String hintText;
-  final Widget resultCardConfirmIcon;
-  final Alignment resultCardAlignment;
-  final Decoration resultCardDecoration;
-  final EdgeInsets resultCardPadding;
+  final Color? appBarColor;
+  final BoxDecoration? searchBarBoxDecoration;
+  final String? hintText;
+  final Widget? resultCardConfirmIcon;
+  final Alignment? resultCardAlignment;
+  final Decoration? resultCardDecoration;
+  final EdgeInsets? resultCardPadding;
 
-  final String language;
+  final String? language;
 
-  final LocationAccuracy desiredAccuracy;
+  final LocationAccuracy? desiredAccuracy;
 
   @override
   LocationPickerState createState() => LocationPickerState();
@@ -73,12 +73,12 @@ class LocationPicker extends StatefulWidget {
 
 class LocationPickerState extends State<LocationPicker> {
   /// Result returned after user completes selection
-  LocationResult locationResult;
+  LocationResult? locationResult;
 
   /// Overlay to display autocomplete suggestions
-  OverlayEntry overlayEntry;
+  OverlayEntry? overlayEntry;
 
-  List<NearbyPlace> nearbyPlaces = List();
+  List<NearbyPlace> nearbyPlaces = [];
 
   /// Session token required for autocomplete API call
   String sessionToken = Uuid().generateV4();
@@ -94,7 +94,7 @@ class LocationPickerState extends State<LocationPicker> {
   /// Hides the autocomplete overlay
   void clearOverlay() {
     if (overlayEntry != null) {
-      overlayEntry.remove();
+      overlayEntry!.remove();
       overlayEntry = null;
     }
   }
@@ -112,14 +112,15 @@ class LocationPickerState extends State<LocationPicker> {
 
     if (place.length < 1) return;
 
-    final RenderBox renderBox = context.findRenderObject();
+    final RenderBox renderBox = context.findRenderObject() as RenderBox;
     Size size = renderBox.size;
 
-    final RenderBox appBarBox = appBarKey.currentContext.findRenderObject();
+    final RenderBox? appBarBox =
+        appBarKey.currentContext!.findRenderObject() as RenderBox?;
 
     overlayEntry = OverlayEntry(
       builder: (context) => Positioned(
-        top: appBarBox.size.height,
+        top: appBarBox!.size.height,
         width: size.width,
         child: Material(
           elevation: 1,
@@ -146,7 +147,7 @@ class LocationPickerState extends State<LocationPicker> {
       ),
     );
 
-    Overlay.of(context).insert(overlayEntry);
+    Overlay.of(context)!.insert(overlayEntry!);
 
     autoCompleteSearch(place);
   }
@@ -159,7 +160,7 @@ class LocationPickerState extends State<LocationPicker> {
 
     // Currently, you can use components to filter by up to 5 countries. from https://developers.google.com/places/web-service/autocomplete
     String regionParam = countries?.isNotEmpty == true
-        ? "&components=country:${countries.sublist(0, min(countries.length, 5)).join('|country:')}"
+        ? "&components=country:${countries!.sublist(0, min(countries.length, 5)).join('|country:')}"
         : "";
 
     var endpoint =
@@ -169,12 +170,13 @@ class LocationPickerState extends State<LocationPicker> {
             "language=${widget.language}";
 
     if (locationResult != null) {
-      endpoint += "&location=${locationResult.latLng.latitude}," +
-          "${locationResult.latLng.longitude}";
+      endpoint += "&location=${locationResult!.latLng!.latitude}," +
+          "${locationResult!.latLng!.longitude}";
     }
 
     LocationUtils.getAppHeaders()
-        .then((headers) => http.get(Uri.parse(endpoint), headers: headers))
+        .then((headers) => http.get(Uri.parse(endpoint),
+            headers: headers as Map<String, String>?))
         .then((response) {
       if (response.statusCode == 200) {
         Map<String, dynamic> data = jsonDecode(response.body);
@@ -214,7 +216,7 @@ class LocationPickerState extends State<LocationPicker> {
   /// To navigate to the selected place from the autocomplete list to the map,
   /// the lat,lng is required. This method fetches the lat,lng of the place and
   /// proceeds to moving the map to that location.
-  void decodeAndSelectPlace(String placeId) {
+  void decodeAndSelectPlace(String? placeId) {
     clearOverlay();
 
     final endpoint =
@@ -223,7 +225,8 @@ class LocationPickerState extends State<LocationPicker> {
             '&language=${widget.language}';
 
     LocationUtils.getAppHeaders()
-        .then((headers) => http.get(Uri.parse(endpoint), headers: headers))
+        .then((headers) => http.get(Uri.parse(endpoint),
+            headers: headers as Map<String, String>?))
         .then((response) {
       if (response.statusCode == 200) {
         Map<String, dynamic> location =
@@ -240,17 +243,18 @@ class LocationPickerState extends State<LocationPicker> {
 
   /// Display autocomplete suggestions with the overlay.
   void displayAutoCompleteSuggestions(List<RichSuggestion> suggestions) {
-    final RenderBox renderBox = context.findRenderObject();
+    final RenderBox renderBox = context.findRenderObject() as RenderBox;
     Size size = renderBox.size;
 
-    final RenderBox appBarBox = appBarKey.currentContext.findRenderObject();
+    final RenderBox? appBarBox =
+        appBarKey.currentContext!.findRenderObject() as RenderBox?;
 
     clearOverlay();
 
     overlayEntry = OverlayEntry(
       builder: (context) => Positioned(
         width: size.width,
-        top: appBarBox.size.height,
+        top: appBarBox!.size.height,
         child: Material(
           elevation: 1,
           child: Column(
@@ -260,7 +264,7 @@ class LocationPickerState extends State<LocationPicker> {
       ),
     );
 
-    Overlay.of(context).insert(overlayEntry);
+    Overlay.of(context)!.insert(overlayEntry!);
   }
 
   /// Utility function to get clean readable name of a location. First checks
@@ -292,7 +296,8 @@ class LocationPickerState extends State<LocationPicker> {
               "location=${latLng.latitude},${latLng.longitude}&radius=150" +
               "&language=${widget.language}";
 
-      return http.get(Uri.parse(endpoint), headers: headers);
+      return http.get(Uri.parse(endpoint),
+          headers: headers as Map<String, String>?);
     }).then((response) {
       if (response.statusCode == 200) {
         nearbyPlaces.clear();
@@ -330,21 +335,24 @@ class LocationPickerState extends State<LocationPicker> {
             "&language=${widget.language}";
 
     final response = await http.get(Uri.parse(endpoint),
-        headers: await LocationUtils.getAppHeaders());
+        headers: await (LocationUtils.getAppHeaders()
+            as FutureOr<Map<String, String>?>));
 
     if (response.statusCode == 200) {
       Map<String, dynamic> responseJson = jsonDecode(response.body);
 
-      String road;
+      String? formattedAddress;
+      List<Map<String, dynamic>>? addressComponents;
+      //see https://developers.google.com/maps/documentation/geocoding/start
 
-      String placeId = responseJson['results'][0]['place_id'];
+      String? placeId = responseJson['results'][0]['place_id'];
 
       if (responseJson['status'] == 'REQUEST_DENIED') {
-        road = 'REQUEST DENIED = please see log for more details';
+        formattedAddress = 'REQUEST DENIED = please see log for more details';
         print(responseJson['error_message']);
       } else {
-        road =
-            responseJson['results'][0]['address_components'][0]['short_name'];
+        addressComponents = responseJson['results'][0]['address_components'];
+        formattedAddress = responseJson['results'][0]['formatted_address'];
       }
 
 //      String locality =
@@ -352,9 +360,45 @@ class LocationPickerState extends State<LocationPicker> {
 
       setState(() {
         locationResult = LocationResult();
-        locationResult.address = road;
-        locationResult.latLng = latLng;
-        locationResult.placeId = placeId;
+        locationResult!.formattedAddress = formattedAddress;
+        locationResult!.latLng = latLng;
+        locationResult!.placeId = placeId;
+        if (addressComponents != null) {
+          addressComponents.forEach((item) {
+            List<String>? types = item['types'];
+            if (types != null && types.length > 0) {
+              switch (types[0]) {
+                case 'street_number':
+                  locationResult!.streetNumber = item['long_name'];
+                  break;
+                case 'route':
+                  locationResult!.route = item['long_name'];
+                  break;
+                case 'sublocality':
+                case 'sublocality_level_1':
+                  locationResult!.subLocality = item['long_name'];
+                  break;
+                case 'locality':
+                  locationResult!.locality = item['long_name'];
+                  break;
+                case 'administrative_area_level_1':
+                  locationResult!.administrativeAreaLevel1 = item['long_name'];
+                  break;
+                case 'administrative_area_level_2':
+                  locationResult!.administrativeAreaLevel2 = item['long_name'];
+                  break;
+                case 'country':
+                  locationResult!.country = item['long_name'];
+                  break;
+                case 'postal_code':
+                  locationResult!.postalCode = item['long_name'];
+                  break;
+                default:
+                  break;
+              }
+            }
+          });
+        }
       });
     }
   }
@@ -362,7 +406,7 @@ class LocationPickerState extends State<LocationPicker> {
   /// Moves the camera to the provided location and updates other UI features to
   /// match the location.
   void moveToLocation(LatLng latLng) {
-    mapKey.currentState.mapController.future.then((controller) {
+    mapKey.currentState!.mapController.future.then((controller) {
       controller.animateCamera(
         CameraUpdate.newCameraPosition(
           CameraPosition(
@@ -373,9 +417,9 @@ class LocationPickerState extends State<LocationPicker> {
       );
     });
 
-    reverseGeocodeLatLng(latLng);
+    //reverseGeocodeLatLng(latLng); //db 13/03/22 : method interesting but not used so create Google charges for nothing
 
-    getNearbyPlaces(latLng);
+    //getNearbyPlaces(latLng);//db 13/03/22 : method interesting but not used so create Google charges for nothing
   }
 
   @override
@@ -442,24 +486,24 @@ class LocationPickerState extends State<LocationPicker> {
 /// set [automaticallyAnimateToCurrentLocation] to false.
 ///
 ///
-Future<LocationResult> showLocationPicker(
+Future<LocationResult?> showLocationPicker(
   BuildContext context,
   String apiKey, {
-  LatLng initialCenter = const LatLng(45.521563, -122.677433),
+  LatLng initialCenter = const LatLng(48.81573732077871, 2.277692232201326),
   double initialZoom = 16,
   bool requiredGPS = false,
-  List<String> countries,
+  List<String>? countries,
   bool myLocationButtonEnabled = false,
   bool layersButtonEnabled = false,
   bool automaticallyAnimateToCurrentLocation = true,
-  String mapStylePath,
+  String? mapStylePath,
   Color appBarColor = Colors.transparent,
-  BoxDecoration searchBarBoxDecoration,
-  String hintText,
-  Widget resultCardConfirmIcon,
-  AlignmentGeometry resultCardAlignment,
-  EdgeInsetsGeometry resultCardPadding,
-  Decoration resultCardDecoration,
+  BoxDecoration? searchBarBoxDecoration,
+  String? hintText,
+  Widget? resultCardConfirmIcon,
+  AlignmentGeometry? resultCardAlignment,
+  EdgeInsetsGeometry? resultCardPadding,
+  Decoration? resultCardDecoration,
   String language = 'en',
   LocationAccuracy desiredAccuracy = LocationAccuracy.best,
 }) async {
@@ -481,8 +525,8 @@ Future<LocationResult> showLocationPicker(
           hintText: hintText,
           searchBarBoxDecoration: searchBarBoxDecoration,
           resultCardConfirmIcon: resultCardConfirmIcon,
-          resultCardAlignment: resultCardAlignment,
-          resultCardPadding: resultCardPadding,
+          resultCardAlignment: resultCardAlignment as Alignment?,
+          resultCardPadding: resultCardPadding as EdgeInsets?,
           resultCardDecoration: resultCardDecoration,
           countries: countries,
           language: language,
