@@ -5,18 +5,39 @@ class LocationResult {
   /// The human readable name of the location. This is primarily the
   /// name of the road. But in cases where the place was selected from Nearby
   /// places list, we use the <b>name</b> provided on the list item.
-  String address; // or road
+  String? formattedAddress; // or road
 
   /// Google Maps place ID
-  String placeId;
+  String? placeId;
 
   /// Latitude/Longitude of the selected location.
-  LatLng latLng;
+  LatLng? latLng;
 
-  LocationResult({this.latLng, this.address, this.placeId});
+  ///streetNumber,route,locality,administrativeAreaLevel2,administrativeAreaLevel1,country of the selected location.
+  String? streetNumber;
+  String? route;
+  String? locality;
+  String? subLocality;
+  String? administrativeAreaLevel2;
+  String? administrativeAreaLevel1;
+  String? country;
+  String? postalCode;
+
+  LocationResult(
+      {this.latLng,
+      this.formattedAddress,
+      this.placeId,
+      this.streetNumber,
+      this.route,
+      this.locality,
+      this.subLocality,
+      this.administrativeAreaLevel2,
+      this.administrativeAreaLevel1,
+      this.country,
+      this.postalCode});
 
   @override
   String toString() {
-    return 'LocationResult{address: $address, latLng: $latLng, placeId: $placeId}';
+    return 'LocationResult{formattedAddress: $formattedAddress, latLng: $latLng, placeId: $placeId, streetNumber: $streetNumber, route: $route, locality: $locality,subLocality $subLocality, administrativeAreaLevel2: $administrativeAreaLevel2, administrativeAreaLevel1: $administrativeAreaLevel1, country: $country, postalCode: $postalCode}';
   }
 }
